changed meaning of components to now refer to all entities; thus requiring a wrapping folder (src). Further cleanup.
