+++ src/app.js
@@ -4,7 +4,8 @@ const config = require('config');
 const swaggerUi = require('swagger-ui-express');
 const fs = require('fs');
 const bodyParser = require('body-parser');
-const swaggerDocument = require('./swagger/swagger.json');
+const swaggerDocument = require('../swagger/swagger.json');
+const routes = require('./components/routes');
 
 const app = express();
 /**
@@ -24,24 +25,23 @@ app.use(bodyParser.json());
 
 app.get('/', (req, res) => res.send('Hello World!'));
 
-// Add components folder to the app.
-app.use(require('./components'));
+// Add routes to the app.
+app.use(routes());
 
 // Swagger for documenting the api, access through localhost:xxxx/api-docs.
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 // Create server with https and cert/key configured.
-const server = https
-  .createServer({
-    key: fs.readFileSync(SERVER_KEY),
-    cert: fs.readFileSync(SERVER_CERT),
-    requestCert: true,
-    rejectUnauthorized: false,
-  }, app)
+const server = https.createServer({
+  key: fs.readFileSync(SERVER_KEY),
+  cert: fs.readFileSync(SERVER_CERT),
+  requestCert: true,
+  rejectUnauthorized: false,
+}, app);
 
 // Listen on port specfied in env-file.
-  .listen(SERVER_PORT,
-    () => console.log(`Example app listening on port ${SERVER_PORT}!`)); // eslint-disable-line no-console
+server.listen(SERVER_PORT,
+  () => console.log(`Example app listening on port ${SERVER_PORT}!`)); // eslint-disable-line no-console
 
 // Export server to use it in tests.
 module.exports = server;
